Add separation of card effect functionality
This commit separates the logic for the Smithy, Adventurer, Sea Hag, Steward, and Great Hall cards from the `cardEffect` function to their own functions.

This commit also adds changes to the implemention of the Smithy, Adventurer, Sea Hag, and Great Hall functions to introduce bugs in the code per the assignment requirements.